move egui wgpu setup to its own module
